requirements.txt
numpy>=1.24
PyYAML>=6.0
Pillow>=10.0
mss>=9.0
opencv-python>=4.8
pygetwindow>=0.0.9
pyautogui>=0.9.54
pytesseract>=0.3.10
